+++ app.js
@@ -1,55 +1,56 @@
-const express = require('express');
-const session = require('express-session');
-const exphbs = require('express-handlebars');
-const path = require('path');
-require('dotenv').config();
+// const express = require('express');
+// const session = require('express-session');
+// const exphbs = require('express-handlebars');
+// const path = require('path');
+// require('dotenv').config();
 
-const bootstrap = require('bootstrap');
-const { sequelize } = require('./models');
+// const { sequelize } = require('./models');
 
-const app = express();
-const PORT = process.env.PORT || 3001;
+// const app = express();
+// // const PORT = process.env.PORT || 3002;
 
-// Middleware to parse JSON and URL-encoded bodies
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
+// const hbs = exphbs.create({});
 
-// Middleware to serve static files from the 'public' directory
-app.use(express.static(path.join(__dirname, 'public')));
+// // Middleware to parse JSON and URL-encoded bodies
+// app.use(express.json());
+// app.use(express.urlencoded({ extended: true }));
 
-// Session management middleware
-app.use(session({
-  secret: process.env.SESSION_SECRET,
-  resave: false,
-  saveUninitialized: false,
-}));
+// // Middleware to serve static files from the 'public' directory
+// app.use(express.static(path.join(__dirname, 'public')));
 
-// Set up Handlebars.js as the templating engine
-app.engine('handlebars', exphbs({ defaultLayout: 'main' }));
-app.set('view engine', 'handlebars');
+// // Session management middleware
+// app.use(session({
+//   secret: process.env.SESSION_SECRET,
+//   resave: false,
+//   saveUninitialized: false,
+// }));
 
-// Import routes
-const authRoutes = require('./routes/authRoutes');
-const gameRoutes = require('./routes/gameRoutes');
-const leaderboardRoutes = require('./routes/leaderboardRoutes');
-const userRoutes = require('./routes/userRoutes');
+// // Set up Handlebars.js as the templating engine
+// app.engine('handlebars', hbs.engine);
+// app.set('view engine', 'handlebars');
 
-// Use routes
-app.use('/auth', authRoutes);
-app.use('/games', gameRoutes);
-app.use('/leaderboard', leaderboardRoutes);
-app.use('/users', userRoutes); // Add user routes
+// // Import routes
+// const authRoutes = require('./routes/authRoutes');
+// const gameRoutes = require('./routes/gameRoutes');
+// const leaderboardRoutes = require('./routes/leaderboardRoutes');
+// const userRoutes = require('./routes/userRoutes');
 
-// Home route
-app.get('/', (req, res) => {
-  res.redirect('/auth/login');
-});
+// // Use routes
+// app.use('/auth', authRoutes);
+// app.use('/games', gameRoutes);
+// app.use('/leaderboard', leaderboardRoutes);
+// app.use('/users', userRoutes); // Add user routes
 
-// Error handling middleware
-const errorHandler = (err, req, res, next) => {
-  console.error(err.stack);
-  res.status(500).send('Something broke!');
-};
-app.use(errorHandler);
+// // Home route
+// app.get('/', (req, res) => {
+//   res.redirect('/auth/login');
+// });
 
-module.exports = app;
+// // Error handling middleware
+// const errorHandler = (err, req, res, next) => {
+//   console.error(err.stack);
+//   res.status(500).send('Something broke!');
+// };
+// app.use(errorHandler);
+
+// module.exports = app;
