--- env_example.txt
@@ -1,3 +1,0 @@
-DB_NAME='office_break_room_db'
-DB_USER=''
-DB_PASSWORD=''
