+++ routes/authRoutes.js
@@ -1,24 +1,24 @@
 const express = require('express');
-const { register, login, logout } = require('../controllers/authController');
+// const { register, login, logout } = require('../controllers/authController');
 
 const router = express.Router();
 
-// Register route
-router.post('/register', register);
+// // Register route
+// router.post('/register', register);
 
-// Login route
-router.post('/login', login);
+// // Login route
+// router.post('/login', login);
 
-// Logout route
-router.get('/logout', logout);
-
-
-// Route to render the login page
-router.get('/login', (req, res) => res.render('auth/login'));
+// // Logout route
+// router.get('/logout', logout);
 
 
-// Route to render the registration page
-router.get('/register', (req, res) => res.render('auth/signup'));
+// // Route to render the login page
+// router.get('/login', (req, res) => res.render('auth/login'));
+
+
+// // Route to render the registration page
+// router.get('/register', (req, res) => res.render('auth/signup'));
 
 
 module.exports = router;
