+++ routes/gameRoutes.js
@@ -4,13 +4,13 @@ const { isAuthenticated } = require('../config/middleware');
 
 const router = express.Router();
 
-// Get all games route
-router.get('/', isAuthenticated, getGames);
+// // Get all games route
+// router.get('/', isAuthenticated, getGames);
 
-// Start game route
-router.get('/play/:id', isAuthenticated, startGame);
+// // Start game route
+// router.get('/play/:id', isAuthenticated, startGame);
 
-// Submit score route
-router.post('/score', isAuthenticated, submitScore);
+// // Submit score route
+// router.post('/score', isAuthenticated, submitScore);
 
 module.exports = router;
