+++ views/dashboard.handlebars
@@ -1,0 +1,29 @@
+
+<h1>Hello</h1>
+
+<p>Test paragraph</p>
+
+<ul>
+    <li><a href="/auth/login">Login</a></li>
+    <li><a href="/auth/register">Register</a></li>
+    <li><a href="/games">Games</a></li>
+    <li><a href="/leaderboard">Leaderboard</a></li>
+    <li><a href="/users">Users</a></li>
+</ul>
+
+
+
+
+
+
+
+
+
+
+{{!-- {{#each user}}
+  <div class="user">
+    <h2>Take a break {{user.name}}</h2>
+    <p>Here are some fun things to do:</p>
+  
+  </div>
+{{/each}} --}}
