--- views/game/mazeGame.handlebars
@@ -1,1 +1,0 @@
-//iframe to display the maze game
