--- views/game/snakeGame.handlebars
@@ -1,1 +1,0 @@
-//iframe to display the snake game
