+++ views/layouts/main.handlebars
@@ -1,4 +1,33 @@
 <!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>{{title}}</title>
+</head>
+<body>
+    {{> header}}
+    <div class="container">
+        {{{body}}}
+    </div>
+    
+</body>
+</html>
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+{{!-- <!DOCTYPE html>
 <html>
   <head>
     <meta charset="utf-8">
@@ -7,32 +36,28 @@
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
     <link rel="stylesheet" href="/css/style.css">
   </head>
-    <body>
-        <header>
-        <h1>Welcome to your Break Room</h1>
-        <h2>A place to hardly work, because you work hard</h2>
-    </header>
-    <nav>
-    <a href="/">Dashboard</a>
-    {{#if loggedIn}}
-    <button id="logout" class="btn-no-style">Logout</button>
-    {{else}}
-    <a href="/login">Login</a>
-    {{/if}}
-  </nav>
-  <main>
-    <aside>
-      <h2>Enjoy the your break!</h2>
-    </aside>
-    <div class="dashboard">
-      {{{body}}}
-    </div>
-  </main>
-  {{#if loggedIn}}
-  <script src="/js/logout.js"></script>
-  {{/if}}
-    
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
-  </body>
+   
+<body>
+  <header>
 
-</html>
+  </header>
+<h1>Welcome to your Break Room</h1>
+    <main>
+      {{!-- {{#if loggedIn}}
+        <div class="container container-fluid mt-5"> --}}
+        {{{body}}}
+        {{!-- </div>
+      {{else}}
+        <div class="container container-fluid mt-5">
+          <h1 class="text-center">log in to access your Office Break Room</h1>
+        </div>
+      {{/if}}
+     --}}
+    </main>
+  <footer>
+
+  </footer>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
+ </body>
+
+</html> --}}
