--- views/layouts/play.handlebars
@@ -1,3 +1,0 @@
-//second page, where the game is played
-//will have general layout to hold game title, game, and game instructions
-//iframe can go here, will link to the game handlebars files
